Refactor out graph internal code too
