+++ new_data_processing/preprocess.py
@@ -4,6 +4,7 @@ from pathlib import Path
 
 import tqdm
 
+from new_data_processing.extract_graph import get_graph
 from new_data_processing.extract_slices import get_slices
 
 logger = logging.getLogger(__name__)
@@ -11,61 +12,6 @@ logger = logging.getLogger(__name__)
 import clang_stuff
 
 from utils import read_csv
-
-
-def extract_nodes_with_location_info(nodes):
-    # Will return an array identifying the indices of those nodes in nodes array,
-    # another array identifying the node_id of those nodes
-    # another array indicating the line numbers
-    # all 3 return arrays should have same length indicating 1-to-1 matching.
-    node_indices = []
-    node_ids = []
-    line_numbers = []
-    node_id_to_line_number = {}
-    for node_index, node in enumerate(nodes):
-        assert isinstance(node, dict)
-        if 'location' in node.keys():
-            location = node['location']
-            if location == '':
-                continue
-            line_num = int(location.split(':')[0])
-            node_id = node['key'].strip()
-            node_indices.append(node_index)
-            node_ids.append(node_id)
-            line_numbers.append(line_num)
-            node_id_to_line_number[node_id] = line_num
-    return node_indices, node_ids, line_numbers, node_id_to_line_number
-    pass
-
-
-def create_adjacency_list(line_numbers, node_id_to_line_numbers, edges, data_dependency_only=False):
-    adjacency_list = {}
-    for ln in set(line_numbers):
-        adjacency_list[ln] = [set(), set()]
-    for edge in edges:
-        edge_type = edge['type'].strip()
-        if True:  # edge_type in ['IS_AST_PARENT', 'FLOWS_TO']:
-            start_node_id = edge['start'].strip()
-            end_node_id = edge['end'].strip()
-            if start_node_id not in node_id_to_line_numbers.keys() or end_node_id not in node_id_to_line_numbers.keys():
-                continue
-            start_ln = node_id_to_line_numbers[start_node_id]
-            end_ln = node_id_to_line_numbers[end_node_id]
-            if not data_dependency_only:
-                if edge_type == 'CONTROLS':  # Control Flow edges
-                    adjacency_list[start_ln][0].add(end_ln)
-            if edge_type == 'REACHES':  # Data Flow edges
-                adjacency_list[start_ln][1].add(end_ln)
-    return adjacency_list
-
-
-def combine_control_and_data_adjacents(adjacency_list):
-    cgraph = {}
-    for ln in adjacency_list:
-        cgraph[ln] = set()
-        cgraph[ln] = cgraph[ln].union(adjacency_list[ln][0])
-        cgraph[ln] = cgraph[ln].union(adjacency_list[ln][1])
-    return cgraph
 
 
 def preprocess(input_data, input_dir, output_dir, project, store=False, vuld_syse=False):
@@ -106,10 +52,7 @@ def preprocess(input_data, input_dir, output_dir, project, store=False, vuld_sys
             logger.debug(f'Skipping node {i} ({file_name}) because len(nodes) == 0')
             continue
 
-        # Make graph data structure
-        node_indices, node_ids, line_numbers, node_id_to_ln = extract_nodes_with_location_info(nodes)
-        adjacency_list = create_adjacency_list(line_numbers, node_id_to_ln, edges, False)
-        combined_graph = combine_control_and_data_adjacents(adjacency_list)
+        combined_graph = get_graph(nodes, edges)
 
         # Tokenize code
         code_text = ' '.join(code.splitlines(keepends=True))
